Handle restarts of the service correctly
Arrange to reconnect out WS if it is closed for any reason. Create a
sequence which returns the current WS, making the subscriber wait if we
are in the process of reconnection.

Change .request to use switchMap; this means we will remake the current
request (with a new UUID) when we are informed of a new WS. The client
must now expect 201s if we reconnect and must understand that these
indicate notifications may have been lost.